refactor : split settings setup wizard step into separate general, mail and registration step components and emit setting updates to parent
